fix: The RSC bundler issues with the Navbar component
