+++ design-system/sections/navbar/navbar.tsx
@@ -3,10 +3,83 @@ import { usePathname } from 'next/navigation'
 import { LINKS } from '@/constants'
 import { localizePath, selectActiveClass } from '@/lib'
 import { useScrollDirection, useScrollProgress } from '@/hooks'
-// import { navbar } from './navbar.styles'
 import { NavbarProps } from './navbar.types'
 import { Logo, ThemeSwitch } from '../../components'
 import { ButtonLink, Nav } from '../../elements'
+import { cva } from '@/styled-system/css'
+import { sharedTransitionProperties } from '../../utils'
+
+// Importing navbar from navbar.styles.tsx creates a circular dependency, so we are using cva here.
+// TODO: Look for a cleaner solution.
+export const navbar = cva({
+  base: {
+    zIndex: 'closest',
+    display: 'flex',
+    justifyContent: 'space-between',
+    width: '100%',
+    paddingX: {
+      base: 'm',
+      sm: 'l',
+      md: 'xl'
+    },
+    paddingY: 'm',
+    backgroundColor: 'gray.900',
+    transitionProperty: 'background-color, border-color, opacity, padding',
+    ...sharedTransitionProperties
+  },
+  variants: {
+    position: {
+      top: {
+        position: 'sticky',
+        top: '0',
+        left: '0'
+      },
+      bottom: {
+        position: 'fixed',
+        bottom: '0',
+        left: '0'
+      }
+    },
+    structure: {
+      nested: {
+        padding: {
+          base: 's',
+          sm: 's',
+          md: '0'
+        },
+        md: {
+          position: 'static',
+          width: 'auto',
+          borderTop: 'none'
+        }
+      }
+    },
+    opacity: {
+      visible: {
+        opacity: '97'
+      },
+      hidden: {
+        opacity: '0',
+        pointerEvents: 'none'
+      }
+    },
+    border: {
+      bottom: {
+        borderBottom: 'gray.subtle'
+      },
+      top: {
+        borderTop: 'gray.subtle'
+      },
+      transparent: {
+        borderColor: 'transparent'
+      }
+    }
+  },
+  defaultVariants: {
+    position: 'top',
+    opacity: 'visible'
+  }
+})
 
 export const Navbar = ({ lang, dictionary }: NavbarProps) => {
   const { links, component, section, layout } = dictionary
@@ -16,10 +89,10 @@ export const Navbar = ({ lang, dictionary }: NavbarProps) => {
 
   return (
     <header
-    // className={navbar({
-    //   opacity: progress > 5 && direction === 'down' ? 'hidden' : 'visible',
-    //   border: progress < 1 ? 'transparent' : 'bottom'
-    // })}
+      className={navbar({
+        opacity: progress > 5 && direction === 'down' ? 'hidden' : 'visible',
+        border: progress < 1 ? 'transparent' : 'bottom'
+      })}
     >
       <Nav
         aria-label={section.navbar.navigation.main}
@@ -34,11 +107,11 @@ export const Navbar = ({ lang, dictionary }: NavbarProps) => {
       </Nav>
       <Nav
         aria-label={section.navbar.navigation.helper}
-        // className={navbar({
-        //   position: 'bottom',
-        //   structure: 'nested',
-        //   border: 'top'
-        // })}
+        className={navbar({
+          position: 'bottom',
+          structure: 'nested',
+          border: 'top'
+        })}
       >
         <ButtonLink
           variant="nav"
